+++ silentphone2/support/axolotl/util/b64helper.h
@@ -20,7 +20,7 @@
  *        be ~1.4 * binLength (rule of thumb).
  * @return number of B64 characters in the @c b64Data buffer.
  */
-int b64Encode(const uint8_t *binData, int32_t binLength, char *b64Data, size_t resultSize);
+size_t b64Encode(const uint8_t *binData, size_t binLength, char *b64Data, size_t resultSize);
 
 /**
  * @brief Decode a Base64 string to binary data
@@ -33,7 +33,7 @@ int b64Encode(const uint8_t *binData, int32_t binLength, char *b64Data, size_t r
  * @param binData Pointer to binary data byte array
  * @return number of binary bytes in the @c binData buffer
  */
-int b64Decode(const char *b64Data, int32_t b64length, uint8_t *binData, size_t binLength);
+size_t b64Decode(const char *b64Data, size_t b64length, uint8_t *binData, size_t binLength);
 
 /**
  * @brief Convert a hex char string to binary array.
@@ -45,7 +45,7 @@ int b64Decode(const char *b64Data, int32_t b64length, uint8_t *binData, size_t b
  * @param target output array
  * @return -1 if an illegals character detected in the string, @c 0 otherwise
  */
-int32_t hex2bin(const char* src, uint8_t* target);
+size_t hex2bin(const char* src, uint8_t* target);
 
 
 /**
